Shutdown programmable clock upon XADC error.
For self-preservation... This is the clock which runs everything except
for the UART interface, so if something goes "wrong" and the XADC
detects that a voltage or temperature is out of spec, this will shutdown
most of the FPGA. Since the serial interface is kept alive, there is a
chance of recovery without reset. It's also possible to query the XADC
to see what might have caused the shutdown.